This commit standardises parameter names and units for some generators, see tac.735).
- `sinusoidal_poisson_generator` and `sinusoidal_gamma_generator` parameters are changed as follows:
  - `dc` -> `rate`
  - `ac` -> `amplitude`
  - `freq` -> `frequency`
  - `phi` -> `phase`; units changed from radian to degrees
- `ppd_sup_generator` parameters are changed as follows:
  - `amplitude` -> `relative_amplitude` (name change only)

This aligns parameters names of the sinusoidal generators with the other generators. The change of parameter name for `ppd_sup_generator` makes explicit that the modulation amplitude is relative to the mean rate for this generator.

All these changes will require adjustment pertaining scripts, they will cause existing code to fail noisily.