Update cppcheck
The cppcheck version that comes along with Ubuntu 14.04 is 1.61.  Apparently, it is a bit buggy, as it halts when executing the line:

    cppcheck --enable=all --inconclusive --std=c++03 sli/tokenutils.cc

The version 1.69 has no such problems, but is not available from the package system. Hence, we fetch it from the git, set it to the correct version and build it ourselves.