Make publicly used dictionaries DictionaryDatums
Before, modeldict_, synapsedict_ and connruledict_ were  handed to the SLIInterpreter in a DictDatum. The DictDatum wants to manage its pointer and does not want any other to free the pointer. If NEST finalizes, it deletes all pointers. Afterwards the SLIInterpreter shuts down and wants to cleanup all DictDatums - bang -> pointer already gone - segfault.

Hence, when we keep publicly used dicts in DictDatums, that manages ref-counting, it will only be freed, when all refs to it have freed it...