Implement former SLI `CheckMultapseAutapseConstraints` ...
in FixedIndegree and FixedOutdegree. As this method was used for special versions of RCC and RDC with only one target or source, respectively, not all checks can efficiently be implemented. In those cases warnings will be emitted.

Also remove the test cases in `ticket-451.sli` that cannot efficiently be tested.