Scaling of the computer epsilon to solve pyNN bug
The following pyNN code produces an issue when running with NEST backend:
>> import pyNN.nest as sim
>> sim.setup(spike_precision='on_grid',timestep=0.1)
>> exc_stim = sim.Population(1, sim.SpikeSourceArray, cellparams={'spike_times': [353538.4]})

By scaling the machine epsilon according to the magnitude of the values being used solve this issue. It is also recommended in the documentation of the epsilon function (http://en.cppreference.com/w/cpp/types/numeric_limits/epsilon).