Reduce the number of build jobs
Many configuration options of NEST are actually independent.
This means we don't have to run the full matrix, but just a
minimal job, a maximal job, and some jobs for dependent options.

This commit simplifies the matrix from 8 runs to 4 runs while
also adding a new option for builds with and without threading.