Add user option to keep/delete source table
The source table is created during connection setup on the postsynaptic
side, but keeps track of presynaptic information. After communication of
connections this can be delete at the expense of increased communication
data for spikes and loss of the ability to create new connections after
the simulation has been started.