Use pandas if available
It tries to load the pandas module and uses it if it can, otherwise, it
falls back to numpy.

Note that the nest.is_iterable test returns True for a string, since a
string is iterable and the function then tries to iterate over each
character in the string! I've, therefore, changed the test.

Tested this out with the gdf file generated from
`examples/nest/tsodyks_shortterm_bursts.sli` using the
`examples/nest/plot_tsodyks_shortterm_bursts.py` plotter script and it
works both with and without pandas.

Fixes #400 to begin with.