Remove all instances of syn_index and use syn_id throughout
Since synapses are not sorted by sources any more as they were in the sparsetable of 4g, we no longer need to distinguish between the syn_id (index in the prototypes array) and syn_index (index in a heterogeneous connector) to reduce memory usage. All heterogeneous connectors (plain vectors in 5g) now have the same size as the prototype array and syn_id can be used to index the correct Connector. Hence, we can remove all occurences of syn_index.