Make the recording backend a per-device setting.
Remove status handling for the recording backends via the kernel and
corresponding Name objects and make this per-device settings. Remove
IOManager::get_/set_recording_backend in favor of explicit functions
and handle enrollment and writing in the RecordingDevice. This also
adds proper handling of /events and /time_in_steps in the recording
backends for memory and screen.