Ensure number of vacant synaptic elements is always >= 0
Minor check to ensure that the number of vacant synaptic elements is
always >= 0. When synaptic elements are to be deleted, z_ <
z_connected_, and in this scenario, this method may return a negative -
which when passed to the `decay_synaptic_elements` method unexpectedly
adds to the z_ value instead of having no effect (there are no
vacant elements to decay).

The decay method is only called after connectivity updates have been
made in the current source, so this issue doesn not arise. However, as
we're thinking of moving towards decoupling the structural plasticity
updates in the future, this general check is worth adding.