Properly set the environment variables for do_tests.sh
This changes the custom build target installcheck to only set the
configuration-specific part of the environment variables PATH and
PYTHONPATH. The bash script do_tests.sh then concatenates these
with the real environment variable. The way it was done before did
not actually work, but was setting the variables to the values at
configure time, instead of build time.

This fixes #817.