Add SymmetricBernoulliConnector
This adds a new connector based on BernoulliConnector, which creates
symmetric connections between nodes with a fixed probability
p. However, instead of iterating over sources for each target and
creating a connection with probability p, this connector draws
a random indegree for each target and then selects number indegree
many random sources. To make sure connections are created
symmetrically also in the distributed case, identical indegrees and
sources need to be drawn on each virtual process. To achieve this, the
connector uses the global rng to seed a local rng on each virtual
process with the same seed.