Correct the handling of zero-delta-t spike pairs
This corrects my previous misunderstanding in the
stdp_nn_pre-centered_synapse model. There, if a postsynaptic
spike occured at the same time as a presynaptic one, that postsynaptic
spike would be discarded, and would never participate in any
potentiation pair, because the nearest pair would be zero-delta-t, and
other pairs would not be looked for.

Now the pairs exactly coinciding (so that presynaptic_spike == postsynaptic_spike
+ dendritic_delay), leading to zero delta_t, are discarded. In this case the
concerned pre/postsynaptic spike is paired with the second latest preceding
post/presynaptic one (for example, pre=={10 ms; 20 ms} and post=={20 ms} will
result in a potentiation pair 20-to-10).