Correct an infinite loop in stdp_nn_restr_synapse
In the logic of stdp_nn_restr_synapse model, when looking for
the earliest postsynaptic spike that occurred after the previous
presynaptic one to account in a potentiation pair, if the time
difference happened to be zero, we discard such a pair but then
take the next postsynaptic spike instead.

Mistakenly, I forgot the command to proceed to the next postsynaptic
spike, thus causing an infinite loop in case a postsynaptic spike
occurs at the same time with a presynaptic one.