Fix parsing of ~/.nestrc in show_help_with_pager()
This fixes the code parsing for the pager command in ~/.nestrc which
could lead to a TypeError exception when calling subprocess.call with
pager=None if ~/.nestrc is an empty file (0 bytes).
In addition, the code now opens ~/.nestrc safely using `with` and uses
only one loop and one regular expression for parsing and uses the last
'/page << /command' entry instead of the first it finds.