Make stdp_facetshw synapse consistent with PR#924
PR #924 fixed direct comparison of delta_t to zero in STDP.
Among other things, that PR added the assertion
  minus_dt < 1.0 * stdp_numerical_imprecision
to all spike-timing-dependent plasticity models. Except
stdp_facetshw synapse, which is fixed in this commit.

In stdp_facetshw synapse the direct comparison to zero was
acceptable, because of the initialization
  double minus_dt = 0;
  double plus_dt = 0;
So, the direct comparison to zero here just checks if `minus_dt`
or `plus_dt` retain their initial values, which basically means
the same as checking `if ( start != finish )`. Still, I remove
the comparison to zero for clarity, since it is redundant.