Introduce time-driven spiking connections
The framework for continuous interactions is used to communicate
spikes. The corresponding synapses are updated in every time
step. While this is much less efficient than the traditional
event-driven approach used in NEST, it allows for fast
prototyping of various plasticity rules that rely on
multiple continuously evolving synaptic state variables. As a
proof of concept the iaf_psc_exp neuron now can be connected via
a new time_driven_static_synapse, which results in identical
dynamics as the corresponding event-driven implementation.