Use an unsigned data type for thread in spike data
thread is typically a signed data type since -1 is used in some cases as a
special value for thread. However to be able to use the full range of thread ids
that are available with NUM_BITS_TID bits in spike data, we need to use an
unsigned data type here.