Getting spatial positions requires passing NodeCollection as argument
Slicing a NodeCollection with spatial metadata would not slice the spatial metadata. The spatial metadata then only reflects the original NodeCollection. So connecting with a sliced NodeCollection would connect using the original un-sliced NodeCollection.

This commit changes how to get positions, requiring that the NodeCollection is passed as an argument. This makes sure that the right nodes are connected, also for sliced NodeCollections.