Remove environment variable NEST_DOC_DIR
This commit removes the possibility to set NEST_DOC_DIR as
this did more harm than good. All uses of the variable have
been replaced by appropriate queries to the NEST kernel. It
also removes the ancient data reading example for SLI, which
used the variable to find an example file.