Update ConnPlotter.py
Forgot one replacement of ".l" by ".left"
The method by which I fixed the "don't use l" warning was
to search the source files for ".l" and replace the instances
with ".left" (manually checking whether .l should be replaced,
or is only a part of a longer variable name starting with .l).
I haven't tested the code yet - are there unit tests?