Warn of SetStatus() between Prepare() and Cleanup() more thoroughly
For me, the warning in the notes in that section was not enough. I changed some things in the simulation code I am working with, including replacing nest.Simulate() by nest.Prepare()/nest.Run()/nest.Cleanup(). After that, there was some SetStatus() between Prepare() and Cleanup(). I didn't catch the note that SetStatus() should not be called between Prepare() and Cleanup(). I continued making other changes, because the code seemed to continue to work. At some point, I realized that there were no/not enough spikes recorded anymore. Tracking down that the problem was in replacing Simulate() by Run() took me on the order of a week, and I only recognized the current warning in the documentation after reverting that change just to see what happens.