Handle axonal delay in original stdp-pl synapse
If the axonal contribution exceeds the dendritic contribution, the results might be erroneous due to relevant post spikes not being available at the time of the synaptic update.
This case triggers a warning but not an error as the model is used as a reference for the new framework supporting long axonal delays.