Variable name correction and adding tests
Some of the parameters that are associated with the fast synaptic
components get '_fast' added to their names.

Pytest script added to test the GLIF model with double alpha synapses.

A couple of test cases require adding extra parameters for the
new glif_psc_double_alpha model, so they are added.